Add HTML frontend to memory_inspector.
This adds the basic infrastructure for the web based ui:
 - A python-based www server.
 - The HTML/JS client.
At the current state, the only functionality available is listing
processes/device stats.

BUG=340294
NOTRY=true

Review URL: https://codereview.chromium.org/190853010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@256646 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c2028b5519fd3bf95c864e6e0c12b1c946b7220f

+++ tools/memory_inspector/memory_inspector/data/file_storage.py
@@ -48,7 +48,8 @@ class Storage(object):
     assert(isinstance(settings, dict))
     file_path = os.path.join(self._root, Storage._SETTINGS_FILE % name)
     if not settings:
-      os.unlink(file_path)
+      if os.path.exists(file_path):
+        os.unlink(file_path)
       return
     with open(file_path, 'w') as f:
       return json.dump(settings, f)
